move uploaded_file_manager.py
