+++ lib/streamlit/caching/hashing.py
@@ -31,7 +31,7 @@ from typing import Any, Pattern, Optional, Dict, List
 from streamlit import type_util
 from streamlit import util
 from streamlit.logger import get_logger
-from streamlit.uploaded_file_manager import UploadedFile
+from streamlit.runtime.uploaded_file_manager import UploadedFile
 from .cache_errors import (
     CacheType,
     UnhashableTypeError,
