+++ lib/tests/streamlit/legacy_caching/hashing_test.py
@@ -53,7 +53,7 @@ from streamlit.legacy_caching.hashing import _CodeHasher
 from streamlit.legacy_caching.hashing import _NP_SIZE_LARGE
 from streamlit.legacy_caching.hashing import _PANDAS_ROWS_LARGE
 from streamlit.type_util import is_type, get_fqn_type
-from streamlit.uploaded_file_manager import UploadedFile, UploadedFileRec
+from streamlit.runtime.uploaded_file_manager import UploadedFile, UploadedFileRec
 import streamlit as st
 
 get_main_script_director = MagicMock(return_value=os.getcwd())
