+++ lib/tests/streamlit/runtime/app_session_test.py
@@ -37,7 +37,7 @@ from streamlit.runtime.scriptrunner import (
 from streamlit.runtime.scriptrunner import ScriptRunnerEvent
 from streamlit.runtime.session_data import SessionData
 from streamlit.state.session_state import SessionState
-from streamlit.uploaded_file_manager import UploadedFileManager
+from streamlit.runtime.uploaded_file_manager import UploadedFileManager
 from streamlit.watcher.local_sources_watcher import LocalSourcesWatcher
 from tests.isolated_asyncio_test_case import IsolatedAsyncioTestCase
 from tests.testutil import patch_config_options
