+++ lib/tests/streamlit/runtime/uploaded_file_manager_test.py
@@ -17,8 +17,8 @@
 import unittest
 
 from streamlit.stats import CacheStat
-from streamlit.uploaded_file_manager import UploadedFileManager
-from streamlit.uploaded_file_manager import UploadedFileRec
+from streamlit.runtime.uploaded_file_manager import UploadedFileManager
+from streamlit.runtime.uploaded_file_manager import UploadedFileRec
 
 FILE_1 = UploadedFileRec(id=0, name="file1", type="type", data=b"file1")
 FILE_2 = UploadedFileRec(id=0, name="file2", type="type", data=b"file222")
