+++ lib/tests/streamlit/web/server/server_test.py
@@ -36,7 +36,7 @@ from streamlit.runtime.forward_msg_cache import ForwardMsgCache
 from streamlit.runtime.forward_msg_cache import populate_hash_if_needed
 from streamlit.logger import get_logger
 from streamlit.proto.ForwardMsg_pb2 import ForwardMsg
-from streamlit.uploaded_file_manager import UploadedFileRec
+from streamlit.runtime.uploaded_file_manager import UploadedFileRec
 from streamlit.watcher import event_based_path_watcher
 from streamlit.web.server.server import DebugHandler
 from streamlit.web.server.server import HealthHandler
