+++ lib/tests/streamlit/web/server/upload_file_request_handler_test.py
@@ -26,7 +26,7 @@ from streamlit.web.server.upload_file_request_handler import (
     UploadFileRequestHandler,
     UPLOAD_FILE_ROUTE,
 )
-from streamlit.uploaded_file_manager import UploadedFileManager
+from streamlit.runtime.uploaded_file_manager import UploadedFileManager
 
 LOGGER = get_logger(__name__)
